Add the basic class and interfaces for load testing
TestCase basically represents a performance test case. It is a runnable and
thus it will be used in a thread. It will keep track of its state, start
time, end time and duration.

TestCaseCreationFactory is responsible for creating test cases.
UnitTestInstance will be used to create the factory.

Added license file and header to the source codes.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@6 71ddd869-3c4c-0410-9474-71dd1778b0a1