Add Observer pattern to engine and test case
Added different listeners and their corresponding events. The purpose of
the state change listeners is to be able to update something else based
on the state. The transition change listener will be implemented based on
the state change listeners and contains methods for particular state of a
test case.

The Load Testing Engine provides API for the Engine to act to perform,
manage and report the test results.

Persistence and UI API design is not final yet, thus their dummy is
created.

All engines will be initialized using a creational pattern probably
factory method pattern.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@7 71ddd869-3c4c-0410-9474-71dd1778b0a1