Make LoadTestEnigne more independent and introduce LoadTestController
LoadTestEngine is now designed and refactored to be responsible only with
test performing behaviors. LoadTestController is introduced to exhibit the
state of the Test; that it is will provide the client with engines
requested.

TestResult is introduced but not designed completely. Its purpose is to
contain the test results. It will be used to communicate between engines
with test related info.

TestCaseThreadPolicy is introduced to provide a way to ask the client to
decide whether to terminate the thread or not.

Added license header to files.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@8 71ddd869-3c4c-0410-9474-71dd1778b0a1