Add assembly of the application and removed duplicate versioning
Assembly the load testing engine is added so that building the application
is done in tendom with project building. Now by simply getting the distrib-
-ution and unzipping it the user can execute the project with one simple
command.

Added the preliminary README file.

Module version was previously explicitly mentioned now they are that of the
parent project. Similar duplicity is avoided for smart dao API as well.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@24 71ddd869-3c4c-0410-9474-71dd1778b0a1