Add abstract persistene implementation which serve as base persistence class
The abstract base persistence engine impl supports the basic operations of
the engine and provides some simple utility methods for the child impls' to
invoke when required.

Add specific value for the progress event.

Change the module description in the license header to be a little more
specific to this module.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@26 71ddd869-3c4c-0410-9474-71dd1778b0a1