Add database script generation and inclusion in distro as part of build
DDLUtils is used to generate and maintain DB scripts. Initially MySQL and
PostgreSQL will be supported for this product and on demand support for
other platforms will be introduced.

Developers should take note and generate db-connection-params.properties
from its provided template. One might also take a look whether any init
script is there for any database (currently there is only for MySQL), if
so execute them first and then set the values of the property. Please note
that this is used by maven.

It generates two sets of SQLs for every database one for the whole system
(suited for fresh installations) and the other (with alter) for upgradation
mainly.

Genarated SQLs are part of the depedencies. Init scripts are not included
as they are for developers and not for fresh installations or upgrades.

Ignored the new to be generated properties file for git and also files
ending with '~'. Often in Linux text generate these files. Committed them
once by mistake so did not want to repeat it.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@44 71ddd869-3c4c-0410-9474-71dd1778b0a1