Add container object ID instead of the object
Object itself is not added to avoid circular reference. Instead the ID is
to be set in the object. Clone now will also clone these integer ids.

In the test case the ID is set to check whether persisted in the API.
Id is set at top to be visible.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@49 71ddd869-3c4c-0410-9474-71dd1778b0a1