Override the equals method to avoid clash in being used in Hash Sets
Before persisting it would be difficult to set an ID for an entity and thus
it could result the object not being in the set in the first place. So
besides the equals method provided by the Smart DAO API custom equals
method is written. Please note if List was used instead of set this
problem would not occur in the first place.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@51 71ddd869-3c4c-0410-9474-71dd1778b0a1