Add Spring IOC to DB Persistent Engine Implementation
Ignore the db config properties file required for testing, it is
actually generated from its .template file. The .template contains
example of what should be available or not.

Hibernate DTD is added in the home directory and also in the
distribution generated as without it hibernate based DAO execution
would not succeed.

POM is changed to include the C3P0 Connection Pooler as it is used
for connection pooling.

Dummy implementation of the DB persistent engines with valid dao
property is added. The dao is basically an concrete class of Smart
Dao.

Spring Application Context XML is added to the resources. It is
currently used by the test methods and later will be used by the
test controller. Current test method only checks whether the
context initializes and delivers beans properly or not. To be
noted that all Dependency Injection is complete.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@56 71ddd869-3c4c-0410-9474-71dd1778b0a1