Fix the clone & isValid implementation
Because of improper cloning the isValid method was not working as
it should as it was not cloning 3 values in TestCaseResult. Now
add their clone lines.

isValid previously in 2 cases did not check for the instance
variable sets in them. So they are added for the validation to
work properly.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@61 71ddd869-3c4c-0410-9474-71dd1778b0a1