Add test case batch creator for creating test case batches
The batch creator will be responsible for creating batches of threads for
test cases which will be started by the load test engine. It is designed to
have one batch per unit test instance.

Client (load test engine) will add observer to it to learn when to retrieve
a batch and start it.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@88 71ddd869-3c4c-0410-9474-71dd1778b0a1