Enhance the SPI to introduce the representation of a batch
Best practice to start, stop multiple threads is through a thread group. In
order to reduce the work of the client creator should create this group.
Event should also encapsulate the Batch so that observers dont have to
make an additional call to get the batch. Instead if the creator is needed
it can be retrieved from the batch.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@97 71ddd869-3c4c-0410-9474-71dd1778b0a1