Fix abstract test case implementation
Previously listener list was enabled after setting the status which is
wrong by convention, i.e. status should be set after operation is performed
and by implementation as set status makes use of the list; thus initialized
it in the first line of the constructor.

Exception should be allowed to throw any exception so that interupt can be
properly detected by the container Runnable.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@105 71ddd869-3c4c-0410-9474-71dd1778b0a1