Make default test case creation factory behave as default
Previously default test case factory used instance member to initialize the
factory which could pose a potential thread safety issue, thus instead of
using the synchronized mechanism, it is scrapped in total in favor of
simple properties based initialization.

Despite being the default factory it was not singleton which is pointless
as it does not use any instance members thus made it singleton.

Earlier it did not register itself as the default creation factory with
default creation factory support now available it is added during
initialization.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@109 71ddd869-3c4c-0410-9474-71dd1778b0a1