Complete engine finish state detection
Add test case life-cycle monitor as until all test cases is finished
running the engine should not die and it also will not know when to make
the test result available.

Batch creator is monitored to determine its end of batch creation as until
no further batch is available one engine can't move to the finished state.

A wrapper for unit test instance is introduced as it is used to wrap a lot
operations required for a test instance, for example maintaining count of
running test cases, storing the respective TestCaseResult to be updated
once a test case finishes implementation and store whether the batch is
done with creating threads.

A separate thread service will be running to determine whether the engine
is in finished state or not. It will be executed if a batch has finished
or if the test case record is finished.

The only thing remaining till now is the integration of test result
generation.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@124 71ddd869-3c4c-0410-9474-71dd1778b0a1