Remove step configuration from API
The just removed step configuration only represents uniform steps while the
system supports uniform and non-uniform, thus it does not belong to the API
in its current form and its suitable form is not required currently. API
users can use the timings to group test case instances and determine test
cases of the same group or timing. It will be re-introduced in proper form
if and only if required.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>


git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@125 71ddd869-3c4c-0410-9474-71dd1778b0a1