Set status related fields on setStatus call at abstract implementation
Setting of start/end date and other related status flag fields (currently
none) should be set from the abstract layer if it supports it and should be
done before the listeners are fired as they might invoke those operations.

This problem was detected by the test cases as it missing start time.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>

git-svn-id: https://smart-load-test.googlecode.com/svn/trunk@129 71ddd869-3c4c-0410-9474-71dd1778b0a1