Fix test result formation
Previously the test case result was not being added to the test result, this
change adds it to the test result when the batch has finished.

Instance number of a test case was not unique across test case, but rather
across the batch only, this change ensures that its unique across test case.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>